- finished load/store instructions

+++ src/decode_utils.c
@@ -2,7 +2,7 @@
 #include <string.h>
 #include "decode.h"
 
-void get_imm_str(Instr *instr_s, uint8_t imm_high, uint16_t imm_low, uint8_t shift, uint8_t positive) {
+void get_imm_str(Instr *instr_s, uint16_t imm_high, uint8_t imm_low, uint8_t shift, uint8_t positive) {
     uint16_t mask = (1 << shift) - 1;
     uint16_t imm = (imm_high << shift) | (imm_low & mask);
     snprintf(instr_s->imm_str, sizeof(instr_s->imm_str), "#%s%d", (positive) ? "" : "-", imm);
